+++ types/src/order.rs
@@ -48,7 +48,7 @@ pub enum OrderStatus {
 // 系统级别的订单
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Order {
-    pub order_id: i64, // 订单ID
+    pub order_id: i32, // 订单ID
     pub strategy_id: i64, // 策略ID
     pub node_id: String, // 节点ID
     pub exchange_order_id: i64, // 交易所订单ID
